Disallow unused and enable clippy's pedantic lints
This is a huge commit, touching almost all files. This was pretty much unavoidable and
splitting it up would have made much more work. This commit essentially only removes the
`#[allow(unused)]` and `#[allow(clippy::pedantic)]` directives from the crate root.

The consequence of this removal is a lot of pedantic error or warnings of unused items/methods.
All of them have been fixed in this single commit and the package now builds without warnings.
The price is the size of this commit and also the existence of issues like #8 and #9. There are
a lot of places where either a `todo!(...)` was added to avoid having to implement some method,
or some attribute/method/... has been annotated with `#[allow(unused)]` to avoid having to make
a decision on whether or not to remove it. Before any release, these cases should be checked.