Remove the dependency on `impl-tools`
We used `autoimpl` provided by the crate `impl-tools` in some places. While this is not really
problematic, it does not really give any benefits over simply implementing traits for references
by ourselves. Certainly, it is easier to parse if the implementation is just in the crate and
not generated by some procedural macro. This commit removes all traces of the dependency.