Remove `...Like` traits
This commit removes traits such as `DFALike`, `DBALike` etc in favor of implementing
methods on the corresponding automaton types.
Additionally, we renamed the `with_colors` method on `TSBuilder` to `with_state_colors`
which more accurately reflects the purpose.
Finally, we refactor some tests to use the builder instead of manually adding states.