Fix completion of transition systems
We fix an issue with the completion of already complete transition
systems. The documentation of this method now test for exactly this edge
case.
In the process, the following further changes were implemented:
- improve the `is_complete` method to use `BTreeSet` instead of
iteration
- move the `is_complete` method from `Determinstic` to
`TransitionSystem`
- add a `with_alphabet_symbols` method to `TSBuilder` for specifying
additional alphabet symbols that do not appear on any transitions

Additionally, we renamed the `collect()` method on `TSBuilder` to
`nondeterministic()`.