Remove `first()`, rework `collect_...` methods and introduce `ForAlphabet<A>`
This commit removes the `first()` method on `IndexType`. To do this, we
also rework all collect methods and introduce a new `ForAlphabet<A>`
trait, which indicates that a type can be constructed for a specific
instance of an alphabet of type `A`.

There are some minor renamings to make everything more consistent. The
`deterministic` and `nondeterministic` methods on `TSBuilder` were
renamed to `into_dts()` and `into_nts()`.
Some collection methods on `Sproutable` were removed and some tests have
been added.