Remove most of the `allow(unused)` directives
A few remain. Some are there to hide implementation details for
different transition systems. Others are in random generation module as
we do not really have a good use for those methods. Finally, there are
also some in the `hoars` submodule because they implement functionality
like implicit edges that is just not used yet.