WIP: seems not so easy
There is a lot of friction when doing this. Every time, a method calls
`with_initial` or similar, this now expects the argument to be wrapped
in `Id(...)`. Similarly, the representation of edges breaks completely.

Best to not do this right now, this will surely make a huge mess and be
of questionable value for the moment.